Item wrap support, emote support, fast as fuck equipping, Apply to all button for itemwrap support

+++ src/routes/mcp/equipbattleroyalecustomization.ts
@@ -17,13 +17,15 @@ const preparedJoinedQueryWithItems = db
     .select()
     .from(profiles)
     .where(and(eq(profiles.type, sql.placeholder('type')), eq(profiles.accountId, sql.placeholder('accountId'))))
-    .leftJoin(items, and(eq(profiles.id, items.profileId), eq(items.id, sql.placeholder('itemToSlot')))).prepare("preparedJoinedQueryOther");
+    .leftJoin(items, and(eq(profiles.id, items.profileId), eq(items.id, sql.placeholder('itemToSlot'))))
+    .leftJoin(attributes, eq(profiles.id, attributes.profileId))
+    .prepare("preparedJoinedQueryOther")
+
+const preparedBoringProfileQuery = db.select().from(profiles).where(and(eq(profiles.type, sql.placeholder('type')), eq(profiles.accountId, sql.placeholder('accountId')))).prepare("preparedBoringProfileQuery");
 
 app.post('/fortnite/api/game/v2/profile/:accountId/client/EquipBattleRoyaleCustomization', async (c) => {
     const t = new Timing("EquipBattleRoyaleCustomization");
-    const validSlots = ["Character", "Backpack", "Pickaxe", "Glider", "SkyDiveContrail", "MusicPack", "LoadingScreen"];
-
-    const notImplementedSlots = ["Dance", "ItemWrap"];
+    const validSlots = ["Character", "Backpack", "Pickaxe", "Glider", "SkyDiveContrail", "MusicPack", "LoadingScreen", "Dance", "ItemWrap"];
 
     const Authorization = c.req.header("Authorization");
     const accountId = getACIDFromJWT(c);
@@ -31,8 +33,6 @@ app.post('/fortnite/api/game/v2/profile/:accountId/client/EquipBattleRoyaleCusto
     const ua = UAParser.parse(c.req.header("User-Agent"));
 
     const body = await Encoding.getJSONBody(c);
-
-    if (notImplementedSlots.includes(body.slotName)) return c.sendError(Aphrodite.mcp.invalidPayload.withMessage("This slot is not implemented yet"));
 
     if (!Authorization) return c.sendError(Aphrodite.authentication.invalidHeader);
     if (!accountId || accountId !== c.req.param("accountId")) return c.sendError(Aphrodite.authentication.notYourAccount);
@@ -52,12 +52,57 @@ app.post('/fortnite/api/game/v2/profile/:accountId/client/EquipBattleRoyaleCusto
     const parsedBody = bodySchema.safeParse(body);
     if (!parsedBody.success) return c.sendError(Aphrodite.mcp.invalidPayload.withMessage(parsedBody.error.errors.map((e) => e.message).join(", ")));
 
-    const { slotName, itemToSlot } = parsedBody.data;
+    const { slotName, itemToSlot, indexWithinSlot } = parsedBody.data;
+
+    let profileChanges = [];
+
+    if (itemToSlot === "" || itemToSlot.includes("_random")) {
+        const [profile] = await preparedBoringProfileQuery.execute({ type: requestedProfileId, accountId });
+        const [attribute] = await db.select().from(attributes).where(and(eq(attributes.profileId, profile.id), eq(attributes.key, `favorite_${slotName.toLowerCase()}`)));
+
+        let valueJSON: string[] | string = attribute ? attribute.valueJSON as string[] : [];
+
+        if (indexWithinSlot === 0) valueJSON = itemToSlot;
+
+        if (["ItemWrap", "Dance"].includes(slotName)) {
+            valueJSON = Array.isArray(valueJSON) ? valueJSON : [];
+            valueJSON[indexWithinSlot] = itemToSlot;
+        }
+
+        const profileChanges = [{
+            changeType: "statModified",
+            name: `favorite_${slotName.toLowerCase()}`,
+            value: valueJSON
+        }];
+
+        Promise.all([
+            bumpRvnNumber.execute({ accountId, type: "athena" }),
+            db.delete(attributes).where(and(eq(attributes.profileId, profile.id), eq(attributes.key, `favorite_${slotName.toLowerCase()}`))),
+            db.insert(attributes).values({
+                profileId: profile.id,
+                key: `favorite_${slotName.toLowerCase()}`,
+                type: requestedProfileId,
+                valueJSON: valueJSON
+            })
+        ]);
+
+        return c.json({
+            profileRevision: profile.revision + 1,
+            profileId: profile.type,
+            profileChangesBaseRevision: profile.revision,
+            profileChanges: profileChanges,
+            profileCommandRevision: profile.revision + 1,
+            serverTime: new Date().toISOString(),
+            multiUpdate: [],
+            responseVersion: 1,
+        });
+    }
 
     const queryResult = await preparedJoinedQueryWithItems.execute({ type: requestedProfileId, accountId, itemToSlot: itemToSlot });
 
     const fetchedProfile = queryResult.map(({ profiles }) => profiles)[0];
     const fetchedItems = queryResult.map(({ items }) => items);
+    const fetchedAttributes = queryResult.map(({ attributes }) => attributes);
 
     const firstItem = fetchedItems[0];
 
@@ -65,22 +110,58 @@ app.post('/fortnite/api/game/v2/profile/:accountId/client/EquipBattleRoyaleCusto
         return c.sendError(Aphrodite.mcp.invalidPayload.withMessage("You do not own the item you are trying to equip"));
     }
 
-    const profileChanges = [{
-        changeType: "statModified",
-        name: `favorite_${slotName.toLowerCase()}`,
-        value: itemToSlot
-    }];
+    if (slotName.toLowerCase() === "dance" || slotName.toLowerCase() === "itemwrap") {
+        const previousAttribute = fetchedAttributes.find((a) => a && a.key === `favorite_${slotName.toLowerCase()}`);
+        let valueJSON: any = previousAttribute ? previousAttribute.valueJSON : [];
 
-    Promise.all([
-        bumpRvnNumber.execute({ accountId, type: "athena" }),
-        db.delete(attributes).where(and(eq(attributes.profileId, fetchedProfile.id), eq(attributes.key, `favorite_${slotName.toLowerCase()}`))),
-        db.insert(attributes).values({
-            profileId: fetchedProfile.id,
-            key: `favorite_${slotName.toLowerCase()}`,
-            type: requestedProfileId,
-            valueJSON: itemToSlot
-        }),
-    ]);
+        if (["ItemWrap", "Dance"].includes(slotName)) {
+            if (!Array.isArray(valueJSON)) {
+                valueJSON = [];
+            }
+            if (indexWithinSlot === -1) {
+                const length = slotName === "ItemWrap" ? 7 : 6;
+                valueJSON = new Array(length).fill(itemToSlot);
+            } else {
+                if (valueJSON.length <= indexWithinSlot) {
+                    valueJSON = valueJSON.concat(new Array(indexWithinSlot - valueJSON.length + 1).fill(""));
+                }
+                valueJSON[indexWithinSlot] = itemToSlot;
+            }
+        }
+
+        Promise.all([
+            bumpRvnNumber.execute({ accountId, type: "athena" }),
+            db.delete(attributes).where(and(eq(attributes.profileId, fetchedProfile.id), eq(attributes.key, `favorite_${slotName.toLowerCase()}`))),
+            db.insert(attributes).values({
+                profileId: fetchedProfile.id,
+                key: `favorite_${slotName.toLowerCase()}`,
+                type: requestedProfileId,
+                valueJSON: valueJSON
+            })
+        ]);
+
+        profileChanges = [{
+            changeType: "statModified",
+            name: slotName.toLowerCase() === "itemwrap" ? "favorite_itemwraps" : "favorite_dance",
+            value: valueJSON
+        }];
+    } else {
+        Promise.all([
+            bumpRvnNumber.execute({ accountId, type: "athena" }),
+            db.insert(attributes).values({
+                profileId: fetchedProfile.id,
+                key: `favorite_${slotName.toLowerCase()}`,
+                type: requestedProfileId,
+                valueJSON: itemToSlot
+            }),
+        ]);
+
+        profileChanges = [{
+            changeType: "statModified",
+            name: `favorite_${slotName.toLowerCase()}`,
+            value: itemToSlot
+        }];
+    }
 
     const rvn = c.req.query("rvn");
     if (!rvn && rvn !== "-1") return c.sendError(Aphrodite.mcp.invalidPayload.withMessage("Missing rvn"));
